+++ src/webrtc.rs
@@ -1,3 +1,4 @@
+pub(crate) mod audio;
 mod factory;
 pub mod p2p_session;
 mod sdp;